docs(NanoLog): :memo: update the NanoLog whiteboard and its exported image
I've enhanced the "Output" section of the example use case of the `%n`
format specifier.
Essentially, I've updated the visual marker, which indicates the
alignment spacing that can be introduced using the `%n` format
specifier, with a different colour and design to make it clearer to the
reader.

As part of this change, I've also updated the whiteboard, and its
exported image.